+++ Utils/CalScene/CalSceneLib/Arguments.h
@@ -20,7 +20,7 @@ namespace NVL_App
 	{
 	private:
 		int _index;
-		int _blockSize;
+		double _blockSize;
 		Vec2i _gridSize;
 		Vec2d _shiftXY;
 		Vec2d _RotYZ;
@@ -33,11 +33,11 @@ namespace NVL_App
 		string _folder;
 
 	public:
-		Arguments(int index, int blockSize, Vec2i& gridSize, Vec2d& shiftXY, Vec2d& RotYZ, double angle, double distance, Vec2d& focals, Point2d& center, Size& imageSize, int decimals, const string& folder) :
+		Arguments(int index, double blockSize, Vec2i& gridSize, Vec2d& shiftXY, Vec2d& RotYZ, double angle, double distance, Vec2d& focals, Point2d& center, Size& imageSize, int decimals, const string& folder) :
 			_index(index), _blockSize(blockSize), _gridSize(gridSize), _shiftXY(shiftXY), _RotYZ(RotYZ), _angle(angle), _distance(distance), _focals(focals), _center(center), _imageSize(imageSize), _decimals(decimals), _folder(folder) {}
 
 		inline int& GetIndex() { return _index; }
-		inline int& GetBlockSize() { return _blockSize; }
+		inline double& GetBlockSize() { return _blockSize; }
 		inline Vec2i& GetGridSize() { return _gridSize; }
 		inline Vec2d& GetShiftXY() { return _shiftXY; }
 		inline Vec2d& GetRotYZ() { return _RotYZ; }
